requirements.txt
gradio
assemblyai
groq
python-dotenv
